projects/market-data-downloader/requirements.txt
ccxt>=4.0.0
requests>=2.31.0
pandas>=2.0.0
pyarrow>=14.0.0
streamlit>=1.30.0
typer>=0.9.0
questionary>=2.0.0
python-dateutil>=2.9.0
